Standardise use of single and double quotation marks
Double quotes are exclusively for quotations. This can include referring to something when it doesn't have its own title, and is instead referred to by a notable line from it (eg. "It is Wednesday, my dudes").

Single quotes are for titles of pieces of media, primarily songs, TV, and movies.